cleaning sockets.py
